+++ src/pages/listInformation/components/SubHeaderBannersLayout/index.ts
@@ -1,0 +1,1 @@
+export { SubHeaderBannersLayout } from './SubHeaderBannersLayout';