+++ CHANGELOG.md
@@ -13,8 +13,10 @@
 * Change save/cancel buttons styles [UILISTS-183]
 * Fix issue with absent blue bar [UILISTS-172]
 * Add cross-tenant logic for list-table [UILISTS-185]
+* Cross-tenant indicator for list detail page [UILISTS-173]
 * Fix GitHub Actions workflow not running for tags [FOLIO-4086]
 
+[UILISTS-173] https://folio-org.atlassian.net/browse/UILISTS-173
 [UILISTS-185] https://folio-org.atlassian.net/browse/UILISTS-185
 [UILISTS-182] https://folio-org.atlassian.net/browse/UILISTS-182
 [UILISTS-183] https://folio-org.atlassian.net/browse/UILISTS-183